requirements.txt
factory_boy
faker
phonenumbers